Split OpenAI completions
