+++ drivers/src/openai/drivers.ts
@@ -1,9 +1,10 @@
 // Combined entry for the concrete OpenAI-protocol drivers, exposed as the `@llumiverse/drivers/openai`
 // subpath so consumers can lazy-load the OpenAI family per-provider without pulling the whole driver
 // barrel (e.g. via `await import('@llumiverse/drivers/openai')`). Kept separate from ./index.ts — which exports the abstract
-// BaseOpenAIDriver that openai.ts/azure_openai.ts/openai_compatible.ts extend — to avoid an ESM
+// BaseOpenAIDriver that openai.ts/azure_openai.ts/openai_comp_responses.ts extend — to avoid an ESM
 // initialization cycle (those files import BaseOpenAIDriver from ./index.js).
 
 export { AzureOpenAIDriver, type AzureOpenAIDriverOptions } from './azure_openai.js';
 export { OpenAIDriver, type OpenAIDriverOptions } from './openai.js';
-export { OpenAICompatibleDriver, type OpenAICompatibleDriverOptions } from './openai_compatible.js';
+export * from './openai_comp_completions.js';
+export { OpenAICompatibleDriver, type OpenAICompatibleDriverOptions } from './openai_comp_responses.js';
